Updating screenshots utitilities

+++ config/rofi/applets/bin/screenshot.sh
@@ -74,26 +74,49 @@ if [[ ! -d "$dir" ]]; then
   mkdir -p "$dir"
 fi
 
-# notify screenshot
-# notify_view() {
-#     dunstify -t 2000 "📸 Screenshot saved"
-# }
 notify_view() {
   local file_path="${dir}/${file}"
+  local action_id="default"
 
-  #     "$file_path"
-  action="$(dunstify --action="open,Open the directory." --action="delete,Delete it." --action="edit,Edit it" -i "$filename" "Screenshot" "Saved & Copied.")"
+  (
 
-  case "$action" in
-    "open")
-      "$TERMINAL" -e sh -c "ranger --selectfile=\"$file_path\"" ;;
-    "delete")
-      rm "$file_path" ;;
-    "edit")
-      setsid --fork pinta "$file_path" ;;
-  esac
-}
+    action="$(dunstify --action="${action_id},Click to view actions" \
+      --timeout=0 \
+      -i "$filename" \
+      "Screenshot" "Saved & Copied.")"
 
+    action="$(echo "$action" | tr -d '[:space:]')"
+
+    if [[ "$action" == "default" ]]; then
+      # Show image in background
+      feh --geometry 700x500 "$file_path" &
+      local feh_pid=$!
+
+      local options="📂\n🗑️\n✏️\n📋"
+
+      local choice="$(echo -e "$options" | rofi -dmenu -p "Screenshot Action" \
+        -theme-str "window {width: 400px;}" \
+        -theme "$theme" \
+        -theme-str "listview {columns: 4; lines: 1;}")"
+
+        # Kill the preview
+        kill $feh_pid 2>/dev/null
+
+        case "$choice" in
+          "📂")
+            nautilus --select "$file_path" ;;
+          "🗑️")
+            rm "$file_path" ;;
+          "✏️")
+            setsid --fork pinta "$file_path" ;;
+          "📋")
+            wl-copy --type image/png < "$file_path" ;;
+        esac
+    fi
+    ) &
+
+    disown
+  }
 
 # Countdown function
 countdown () {
@@ -106,33 +129,28 @@ countdown () {
 # Screenshot functions
 shotnow () {
   hyprshot --mode output --raw > ~/Pictures/Screenshots/"$file" 
-  wl-copy --type image/png < ~/Pictures/Screenshots/"$file" 
   notify_view
 }
 
 shot5 () {
   countdown 5
   sleep 1 && hyprshot --mode output --raw >  ~/Pictures/Screenshots/"$file"
-  wl-copy --type image/png < ~/Pictures/Screenshots/"$file" 
   notify_view
 }
 
 shot10 () {
   countdown 10
   sleep 1 && hyprshot --mode output --raw >  ~/Pictures/Screenshots/"$file"
-  wl-copy --type image/png < ~/Pictures/Screenshots/"$file" 
   notify_view
 }
 
 shotwin () {
   hyprshot --mode window --raw >  ~/Pictures/Screenshots/"$file"
-  wl-copy --type image/png < ~/Pictures/Screenshots/"$file" 
   notify_view
 }
 
 shotarea () {
   hyprshot --mode region --raw >  ~/Pictures/Screenshots/"$file"
-  wl-copy --type image/png < ~/Pictures/Screenshots/"$file" 
   notify_view
 }
